Adds conditionals for bindgen and rename use (#14)
The low-level sys crate will be compiled differently depending on
whether renaming is used in the ICU library and whether bindgen
is invoked.

The earlier version of the library had a clunkier mechanism that handles
these separate configurations.  This version is slightly less clunky,
and unifies the bindgen and not-bindgen approaches.

For the time being, the non-bindgen is only useful for ICU version 64.x.

Fixes #13.