Support non-generated versioned builds.
Adds the ability to build bindings for a user-defined
ICU library version, and to turn off build feature detection.

Since there are already plenty of configurations that one could
build the library in, I don't expect this to be the last fixup here.
But we got to start from somewhere.

Bumps the version to 0.0.4.

Fixes #23