Adds ICU test environments by version (#42)
* Adds ICU test environments by version

This will allow us to run CI builds for each combination of ICU versions
and features.  Issue #38 will be resolved once all environments of
interest are being tested.  This commit adds only one new test target,
as an example.

* Adds tests for ICU versions 64, 65 and 66.

This ensures that all versions of ICU of interest are well tested.

Fixes #38