Starts testing the library with ICU 63
Looking into a resolution of issue #41, seems like we'd need
to modify the build environment of `docs.rs` to include the
development ICU library.  Looking at the packages available,
it seems that this is going to have to be ICU version 63.

Hence, adding the build and test envs with ICU 63.  They will be
linked into the build process as a followup.