Use `bindgen` library to generate bindings
Uses the `bindgen` library, instead of a `bindgen` executable
to generate the library bindings.

This helps in issue #41, working towards a set of crates with
documentation that compiles on `docs.rs`.  Also, it places the
responsiblity for bindgen versioning in the hands of cargo, which
is better than needing to have the correct version installed.