Adds basic collation support
Collation (locale-sensitive sorting) is one of a few
high-value features which lean on *a lot* of the ICU
functionality and CLDR-supplied data in a way that can
not easily be replicated by other solutions.

This commit adds a basic collation API, which covers only
the main use cases of collation.

Fixes: 85