Prepares for continuous feature testing
As we pondered what is needed to start using rust_icu with version 67.1
for Fuchsia, it became apparent that we have not tested the feature set
that includes `rust_icu_version_in_env` that Fuchsia uses.

This change makes a few modifications that will ultimately allow us to
turn such testing on in the CI environment.  See issue #90 for details.

- `entrypoint.sh`: The old code did not exit with an error when
  subprocessed `cargo test` was failing.
- `make docker-test` now forwards the ICU version to be used for
  testing
- Also, forwards the cargo test args (which will be used to switch
  various features off and on).
- Adds `lib_67.rs` which is a bindgen-generated source code based off of
  ICU 67.1, for Fuchsia to use.  We will need to add other versions, as
  well as update old ones, to support newer features of rust_icu.