Fixes a number of mistakes in the feature setup
This commit fixes a number of problems with the way the feature flags
were handled, which made all flavors of build effectively the same.

It turns out that feature-activation in the top-level directory
that has a virtual manifest does nothing.  In fact, newer cargo
versions make it an error to test with features set in that directory.

This in turn means that our thus-far feature testing actually didn't
test non-default features.  Oops.

Fix this by descending manually into each directory and separately
test and generate docs.  This change requires a buildenv re-release to
take effect.

A couple of mishaps here.

- Despite what the documentation says, the features in the `--features`
  arg list can be separated by commas, and not spaces.  Somewhat
  weirdly, `--features foo bar` works, but `--features foo bar baz` is
  an error.  However, `--features foo,bar,baz` works as does
  `--features foo --features bar --features baz`.

  Separation by commas is less confusing in our case, so we use that.

- When `--features foo` is used, then the cfg switch in the source code
  is, in fact, `#[cfg(feature="foo")]`.  However `cargo` permits putting
  anything in `cargo:rust-cfg=...`, and `rust_icu_sys` had
  `cargo:rust-cfg=features="foo"` there.  Note `featureS`, not
  `feature`, there is an extra "s" at the end.  Now, this works just
  fine, *but* it doesn't interact with cargo's `--features` flag.

  The code fixes build.rs to generate `feature="foo"`, and the source
  code is fixed to change `#[cfg(features="foo")]` to
  `#[cfg(feature="foo")]`.

Issue #120.