Moves 'experimental' to 'proposals'
The motivation is that the proposals are not experimental *code*, but
rather, they are documents that are meant to be read, and happen to have
some bits that you can copy over to the rust playground.

So in that respect, the texts are no different than any other doc you
may write, hence the move to proposals.