rust_icu_uloc: Make build.rs no-op if icu_config is disabled
`build.rs` is used in this crate to discover the ICU library configuration
based on the installed ICU tools, when the user requests it (via feature
`icu_config`).

However, the use of ICU tools is not required to configure the crate. The user
can also supply the values manually, which removes the need to have the
auto-configuration tools. In which case we can make do with an empty `build.rs`.