Adds support for UFormattable.
Using uformattable::UFormattable provides
generalized value parsing for the particular
locale.

While this isn't the full spectrum of the
underlying functions that ICU4C provides
for value parsing, it's general enough
to be usable for all the use cases where
parsing is needed, *if* needed.  However,
those uses are rare.

Issue #141.