Makes ecma402_traits an optional dependency
This removes the need for every use of `rust_icu_uloc` to
depend on `ecma420_traits` if the marker implementation of
`ecma402_traits::Locale` is not in fact being used.

This happens for example when you are using `rust_icu`
directly and not through any ECMA402 compatible API. This change
should, for example, remove the need for Fuchsia code base to
vendor ecma402_traits.

The github workflows that specify custom features need this feature
to be enabled in order to compile correctly, so the feature was added
to all workflow jobs.

Closes #171.