build: makes include dirs more precise
Modifies `build.rs` to use the include directory as output by
`pkg-config --cppflags` instead of just the install prefix.

It seems as if this doesn't matter but there are distributions
that take creative liberties with the directory layout and that
confused the old code.

The new code uses the list of include directories, and verifies that
the needed header files are in fact available on the file system
during `build.rs` execution.

See #189