Deflake the date time format test
The date time format test didn't have a time zone chosen.
This meant that the date time formatting was done based
on the local time zone, which can vary across machines.

I didn't catch this originally because I worked on a
machine that was in UTC, just the same as the machines
which were running the tests.