Adopts the C++ comments from the ICU library
Earlier versions of cargo insisted on running comments generated
from C++ code as rust tests when the comment lines were indented.

This is of course not correct behavior for generated code. Since at
the time there was no way to turn the behavior off, we had to turn
off documentation generation.

Now that the behavior seems to be implemented, we can start
generating docs again.