Suppress nullptr dereference warning in rust_icu_sys/src/lib.rs.
The warning comes about because `bindgen` chooses to use an UB idiom to
generate tests. There is nothing that `rust_icu` can reasonably do about
it.  But since `bindgen` folks are interested in resolving the issue, we
can hope that future `bindgen` versions will resolve the issue.

In the meantime, let's not pollute the (already verbose) build output with
these warnings.

See: https://github.com/rust-lang/rust-bindgen/issues/1651 for details.