Use make dependency tracking for some crate publications
Broke the target `publish` into a few smaller targets and added the correct
dependencies between those targets.

This removes a certain class of errors from publishing crates in the wrong
order.  The approach could be continued to its logical conclusion, but for
the time being I opted not to do that.