Fix the buffer preflight treatment and add tests
The character conversion code uses a two-pass approach for converting
potentially large inputs into various Unicode strings.  The first
pass uses a fixed capacity buffer (often 200 units).  For strings
that happen to be shorter, the conversion succeeds and nothing else
needs to be done.

But for longer strings, the conversion fails, returning a non-OK status
and the size of the buffer that is actually needed.  A second pass then
provides a buffer that's exactly long enough.

That's great and all, but `status` is reused across both calls, and when
the conversion fails, `status` needs to be set back to `OK_CODE` before
the 2nd pass is attempted.

Props for having the assertion that checks for this; but boos for not
actually testing the conversion code with long strings.

Fixes: #244