Exposes a few new interesting symbols in rust_icu_sys
Some symbols requested were from the 'tools' repo which is not part of the
released API, so I added a way to generate those too.

Since this is not relevant for the "regular" rust_icu_sys, I added a
separate Makefile target.

Run the following to regenerate:

```
env ICU_SOURCE_DIR=$HOME/code/icu make static-bindgen-special
```

The results and tools are in `./rust_icu/rust_icu_sys/bindgen_special`

While I was here, I added a separate `rustfmt` invocation. This used
to work automatically up to a point with `bindgen`, but then with the
rename of `rustfmt` it stopped.

Issue: 264