feat: corrects the public API for udata_open
I realized that the stringly typed API for a `udata_open` wrapper
leads to confusing API usages on Rust side.  I added some type
improvements to try to fix that.

This required releasing 4.2.2. I think it will be OK since nobody
is using this code yet, and updates are trivial.