feat: create a container for building ICU against `main`
For the longest time, we had no way to run tests of rust_icu
against top-of-tree ICU. This meant that if a build break was
in the works upstream, we would not know until the release.

This change fixes the issue. We now maintain a "current" environment
which will download and build ICU from `main` before running tests.